move bridge into its' own file
