+++ automation/cypress/support/commands.js
@@ -1,103 +1,4 @@
-import { Eip1193Bridge } from "@ethersproject/experimental/lib/eip1193-bridge";
-import { JsonRpcProvider } from "@ethersproject/providers";
-import { Wallet } from "@ethersproject/wallet";
-import { ethers } from "ethers";
-
-export const TEST_PRIVATE_KEY = Cypress.env("INTEGRATION_TEST_PRIVATE_KEY");
-// Address of the above key
-export const TEST_ADDRESS_NEVER_USE = new Wallet(TEST_PRIVATE_KEY).address;
-
-class CustomizedBridge extends Eip1193Bridge {
-  chainId = 3;
-
-  async sendAsync(...args) {
-    console.debug("sendAsync called", ...args);
-    return this.send(...args);
-  }
-  async send(...args) {
-    console.debug("send called", ...args);
-    const isCallbackForm =
-      typeof args[0] === "object" && typeof args[1] === "function";
-    let callback;
-    let method;
-    let params;
-    if (isCallbackForm) {
-      callback = args[1];
-      method = args[0].method;
-      params = args[0].params;
-    } else {
-      method = args[0];
-      params = args[1];
-    }
-    // Mock out request accounts and chainId
-    if (method === "eth_requestAccounts" || method === "eth_accounts") {
-      if (isCallbackForm) {
-        callback({ result: [TEST_ADDRESS_NEVER_USE] });
-      } else {
-        return Promise.resolve([TEST_ADDRESS_NEVER_USE]);
-      }
-    }
-    if (method === "eth_chainId") {
-      if (isCallbackForm) {
-        callback(null, { result: "0x3" });
-      } else {
-        return Promise.resolve("0x3");
-      }
-    }
-    try {
-      // Hacky, https://github.com/ethers-io/ethers.js/issues/1683#issuecomment-1016227588
-
-      // If from is present on eth_call it errors, removing it makes the library set
-      // from as the connected wallet which works fine
-      if (params && params.length && params[0].from && method === "eth_call")
-        delete params[0].from;
-      let result;
-      // For sending a transaction if we call send it will error
-      // as it wants gasLimit in sendTransaction but hexlify sets the property gas
-      // to gasLimit which makes sensd transaction error.
-      // This has taken the code from the super method for sendTransaction and altered
-      // it slightly to make it work with the gas limit issues.
-      if (
-        params &&
-        params.length &&
-        params[0].from &&
-        method === "eth_sendTransaction"
-      ) {
-        // Hexlify will not take gas, must be gasLimit, set this property to be gasLimit
-        params[0].gasLimit = params[0].gas;
-        delete params[0].gas;
-        // If from is present on eth_sendTransaction it errors, removing it makes the library set
-        // from as the connected wallet which works fine
-        delete params[0].from;
-        const req = ethers.providers.JsonRpcProvider.hexlifyTransaction(
-          params[0]
-        );
-        // Hexlify sets the gasLimit property to be gas again and send transaction requires gasLimit
-        req.gasLimit = req.gas;
-        delete req.gas;
-        // Send the transaction
-        const tx = await this.signer.sendTransaction(req);
-        result = tx.hash;
-      } else {
-        // All other transactions the base class works for
-        result = await super.send(method, params);
-      }
-      console.debug("result received", method, params, result);
-      if (isCallbackForm) {
-        callback(null, { result });
-      } else {
-        return result;
-      }
-    } catch (error) {
-      console.log(error);
-      if (isCallbackForm) {
-        callback(error, null);
-      } else {
-        throw error;
-      }
-    }
-  }
-}
+import { createBridge } from "./eip1193-bridge";
 
 // sets up the injected provider to be a mock ethereum provider with the given mnemonic/index
 // eslint-disable-next-line no-undef
@@ -107,12 +8,7 @@ Cypress.Commands.overwrite("visit", (original, url, options) => {
     onBeforeLoad(win) {
       options && options.onBeforeLoad && options.onBeforeLoad(win);
       win.localStorage.clear();
-      const provider = new JsonRpcProvider(
-        `https://ropsten.infura.io/v3/4f846e79e13f44d1b51bbd7ed9edefb8`,
-        3
-      );
-      const signer = new Wallet(TEST_PRIVATE_KEY, provider);
-      win.ethereum = new CustomizedBridge(signer, provider);
+      win.ethereum = createBridge();
     },
   });
 });
